refactor: move nav drawer to separate component
