+++ src/layouts/MainLayout.vue
@@ -1,7 +1,23 @@
 <template>
   <q-layout class="non-selectable column no-wrap" view="hHh LpR lFr">
+    <!-- Topbar -->
     <HeaderBase />
 
+    <!-- Side navigation -->
+    <NavDrawer v-model="miniState" />
+
+    <!-- Main content -->
+    <q-scroll-area
+      :bar-style="barStyle()"
+      :thumb-style="thumbStyle()"
+      style="flex: 1 1 1px"
+    >
+      <q-page-container class="main-bg">
+        <router-view />
+      </q-page-container>
+    </q-scroll-area>
+
+    <!-- Footer -->
     <q-footer
       v-if="
         currentSettings?.enableMediaDisplayButton ||
@@ -14,149 +30,27 @@
     >
       <ActionIsland />
     </q-footer>
-
-    <q-drawer
-      v-model="drawer"
-      :breakpoint="5"
-      :mini="miniState"
-      bordered
-      class="column justify-between no-wrap bg-secondary-contrast text-weight-medium text-dark-grey"
-    >
-      <q-item
-        v-if="$q.screen.gt.xs"
-        v-ripple
-        clickable
-        @click="miniState = !miniState"
-      >
-        <q-tooltip
-          v-if="miniState"
-          :delay="1000"
-          anchor="center right"
-          self="center left"
-        >
-          {{ $t('expand-sidebar') }}
-        </q-tooltip>
-        <q-item-section avatar>
-          <q-icon name="mmm-menu" />
-        </q-item-section>
-        <q-item-section>{{ $t('collapse-sidebar') }}</q-item-section>
-      </q-item>
-      <q-item
-        v-ripple
-        :disable="!currentSettings || invalidSettings()"
-        :to="{ path: '/media-calendar', exact: true }"
-        active-class="bg-accent-100 text-primary blue-bar"
-        clickable
-      >
-        <q-tooltip
-          v-if="miniState"
-          :delay="1000"
-          anchor="center right"
-          self="center left"
-        >
-          {{
-            !currentSettings
-              ? $t('select-a-congregation-to-enable')
-              : $t('titles.meetingMedia')
-          }}
-        </q-tooltip>
-        <q-item-section avatar>
-          <q-icon name="mmm-media" />
-        </q-item-section>
-        <q-item-section>{{ $t('titles.meetingMedia') }}</q-item-section>
-      </q-item>
-      <q-item
-        v-ripple
-        :disable="!currentSettings || invalidSettings() || mediaPlaying"
-        :to="{ path: '/present-website', exact: true }"
-        active-class="bg-accent-100 text-primary blue-bar"
-        clickable
-      >
-        <q-tooltip
-          v-if="miniState"
-          :delay="1000"
-          anchor="center right"
-          self="center left"
-        >
-          {{ $t('titles.presentWebsite') }}
-        </q-tooltip>
-        <q-item-section avatar>
-          <q-icon name="mmm-open-web" />
-        </q-item-section>
-        <q-item-section>{{ $t('titles.presentWebsite') }}</q-item-section>
-      </q-item>
-      <q-item
-        v-ripple
-        :disable="mediaPlaying"
-        :to="{ path: '/congregation-selector', exact: true }"
-        active-class="bg-accent-100 text-primary blue-bar"
-        clickable
-      >
-        <q-tooltip
-          v-if="miniState"
-          :delay="1000"
-          anchor="center right"
-          self="center left"
-        >
-          {{ $t('titles.profileSelection') }}
-        </q-tooltip>
-        <q-item-section avatar>
-          <q-icon name="mmm-groups" />
-        </q-item-section>
-        <q-item-section>
-          {{ $t('titles.profileSelection') }}
-        </q-item-section>
-      </q-item>
-      <q-space />
-      <q-item
-        v-ripple
-        :disable="
-          !currentSettings || mediaPlaying || route.fullPath.includes('wizard')
-        "
-        :to="{ path: '/settings', exact: true }"
-        active-class="bg-accent-100 text-primary blue-bar"
-        clickable
-      >
-        <q-tooltip
-          v-if="miniState"
-          :delay="1000"
-          anchor="center right"
-          self="center left"
-        >
-          {{ $t('titles.settings') }}
-        </q-tooltip>
-        <q-item-section avatar>
-          <q-icon
-            :color="invalidSettings() ? 'negative' : ''"
-            name="mmm-settings"
-          />
-        </q-item-section>
-        <q-item-section :class="invalidSettings() ? 'text-negative' : ''">
-          {{ $t('titles.settings') }}
-        </q-item-section>
-      </q-item>
-    </q-drawer>
-    <q-scroll-area
-      :bar-style="barStyle()"
-      :thumb-style="thumbStyle()"
-      style="flex: 1 1 1px"
-    >
-      <q-page-container class="main-bg">
-        <router-view />
-      </q-page-container>
-    </q-scroll-area>
   </q-layout>
 </template>
 
 <script setup lang="ts">
+// Packages
 import { storeToRefs } from 'pinia';
-import { Dark, LocalStorage } from 'quasar';
-import { useQuasar } from 'quasar';
-import { queues } from 'src/boot/globals';
-import { barStyle, thumbStyle } from 'src/boot/globals';
+import { Dark, LocalStorage, useQuasar } from 'quasar';
+import { onMounted, ref, watch } from 'vue';
+import { useI18n } from 'vue-i18n';
+import { useRoute, useRouter } from 'vue-router';
+
+// Globals
+import { barStyle, queues, thumbStyle } from 'src/boot/globals';
 import { refreshDateLocale } from 'src/boot/i18n';
+
+// Components
 import HeaderBase from 'src/components/header/HeaderBase.vue';
 import ActionIsland from 'src/components/ui/ActionIsland.vue';
+import NavDrawer from 'src/components/ui/NavDrawer.vue';
+
+// Helpers
 import {
   cleanAdditionalMediaFolder,
   cleanLocalStorage,
@@ -170,14 +64,17 @@ import {
   unregisterAllCustomShortcuts,
 } from 'src/helpers/keyboardShortcuts';
 import { createTemporaryNotification } from 'src/helpers/notifications';
+
+// Stores
 import { useAppSettingsStore } from 'src/stores/app-settings';
 import { useCongregationSettingsStore } from 'src/stores/congregation-settings';
 import { useCurrentStateStore } from 'src/stores/current-state';
 import { useJwStore } from 'src/stores/jw';
-import { onMounted, ref, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useRoute, useRouter } from 'vue-router';
 
+// Local state
+const miniState = ref(true);
+
+// Icon mapping
 const $q = useQuasar();
 
 $q.iconMapFn = (iconName) => {
@@ -198,27 +95,33 @@ $q.iconMapFn = (iconName) => {
   };
 };
 
-// Store and router initializations
+// Routes and translations
+const route = useRoute();
+const router = useRouter();
+const { locale, t } = useI18n({ useScope: 'global' });
 
+// Store initializations
 const appSettings = useAppSettingsStore();
 const { migrations } = storeToRefs(appSettings);
 const { runMigration } = appSettings;
+
+if (!migrations.value.includes('firstRun')) {
+  const migrationResult = runMigration('firstRun');
+  if (migrationResult) {
+    createTemporaryNotification({
+      caption: t('successfully-migrated-from-the-previous-version'),
+      icon: 'mmm-info',
+      message: t('welcome-to-mmm'),
+      timeout: 15000,
+      type: 'positive',
+    });
+  }
+}
 
 appSettings.$subscribe((_, state) => {
   LocalStorage.set('migrations', state.migrations);
   LocalStorage.set('screenPreferences', state.screenPreferences);
 });
-
-const currentState = useCurrentStateStore();
-const { invalidSettings } = currentState;
-const {
-  currentCongregation,
-  currentSettings,
-  downloadProgress,
-  mediaPlaying,
-  online,
-  selectedDate,
-} = storeToRefs(currentState);
 
 const congregationSettings = useCongregationSettingsStore();
 congregationSettings.$subscribe((_, state) => {
@@ -226,7 +129,6 @@ congregationSettings.$subscribe((_, state) => {
 });
 
 const jwStore = useJwStore();
-const { updateJwLanguages } = jwStore;
 jwStore.$subscribe((_, state) => {
   LocalStorage.set('jwLanguages', state.jwLanguages);
   LocalStorage.set('jwSongs', state.jwSongs);
@@ -237,24 +139,17 @@ jwStore.$subscribe((_, state) => {
   LocalStorage.set('lookupPeriod', state.lookupPeriod);
 });
 
-const {
-  fs,
-  getAppDataPath,
-  getUserDataPath,
-  getUserDesktopPath,
-  path,
-  readShortcutLink,
-  setAutoStartAtLogin,
-  writeShortcutLink,
-} = electronApi;
-
-const { locale, t } = useI18n({ useScope: 'global' });
-const drawer = ref(true);
+const { updateJwLanguages } = jwStore;
 updateJwLanguages();
 
-const route = useRoute();
-const router = useRouter();
-const miniState = ref(true);
+const currentState = useCurrentStateStore();
+const {
+  currentCongregation,
+  currentSettings,
+  downloadProgress,
+  online,
+  selectedDate,
+} = storeToRefs(currentState);
 
 watch(currentCongregation, (newCongregation, oldCongregation) => {
   try {
@@ -352,6 +247,17 @@ watch(
   },
 );
 
+const {
+  fs,
+  getAppDataPath,
+  getUserDataPath,
+  getUserDesktopPath,
+  path,
+  readShortcutLink,
+  setAutoStartAtLogin,
+  writeShortcutLink,
+} = electronApi;
+
 watch(
   () => currentSettings.value?.autoStartAtLogin,
   (newAutoStartAtLogin) => {
@@ -380,19 +286,6 @@ watch(
     }
   },
 );
-
-if (!migrations.value.includes('firstRun')) {
-  const migrationResult = runMigration('firstRun');
-  if (migrationResult) {
-    createTemporaryNotification({
-      caption: t('successfully-migrated-from-the-previous-version'),
-      icon: 'mmm-info',
-      message: t('welcome-to-mmm'),
-      timeout: 15000,
-      type: 'positive',
-    });
-  }
-}
 
 // Hack for Windows shortcut glitch, possibly related to https://github.com/electron-userland/electron-builder/issues/2435
 try {
@@ -462,15 +355,6 @@ watch(
       setTimeout(() => {
         attemptedClose.value = false;
       }, 10000);
-    }
-  },
-);
-
-watch(
-  () => $q?.screen?.lt?.sm,
-  (isNowExtraSmall) => {
-    if (isNowExtraSmall) {
-      miniState.value = true;
     }
   },
 );
